Added benchmarks and optimizations
Added a linear search index for comparison with the radix tree index
and then benchmarked the two implementations.  The benchmarks revealed
a number of inefficiencies in the code and this commit includes
updates to the radix tree index to improve performance.  At this time
the radix tree search is about 1.8 times faster than linear search,
although radix tree adds are about half as fast as linear adds using
Go's append.